[BUILD][MISC] Rewritten all Makefiles for efficiency and correctness

+++ dev_handler/dev_handler.c
@@ -5,10 +5,10 @@
 
 #include <termios.h>  // for POSIX terminal control definitions in serialport_open()
 
-#include "../logger/logger.h"
-#include "../runtime_util/runtime_util.h"
-#include "../shm_wrapper/shm_wrapper.h"
-#include "message.h"
+#include <dev_handler_message.h>
+#include <logger.h>
+#include <runtime_util.h>
+#include <shm_wrapper.h>
 
 /**
  * Each device will have a unique port number.
@@ -24,7 +24,7 @@
  * These file paths may also be referred to as "port_prefix" in the code.
  */
 #define LOWCAR_FILE_PATH "/dev/ttyACM"
-#define VIRTUAL_FILE_PATH "ttyACM"			// will be created in the home directory
+#define VIRTUAL_FILE_PATH "ttyACM"  // will be created in the home directory
 #define LOWCAR_USB_FILE_PATH "/dev/ttyUSB"
 
 // **************************** PRIVATE STRUCT ****************************** //
@@ -93,7 +93,7 @@ uint32_t used_lowcar_usb_ports = 0;
 pthread_mutex_t used_ports_lock;  // poll_connected_devices() and relay_clean_up() shouldn't access used_ports at the same time
 
 // String to hold the home directory path (for looking for virtual device sockets)
-const char *home_dir;
+const char* home_dir;
 
 #define MAX_PORT_NAME_SIZE 64
 
@@ -395,7 +395,7 @@ void relay_clean_up(relay_t* relay) {
     pthread_mutex_unlock(&used_ports_lock);
     pthread_mutex_destroy(&relay->relay_lock);
     pthread_cond_destroy(&relay->start_cond);
-    if (relay->dev_id.uid == -1) {
+    if (relay->dev_id.uid == (uint64_t) -1) {
         char port_name[MAX_PORT_NAME_SIZE];
         construct_port_name(port_name, relay->is_virtual, relay->is_usb, relay->port_num);
         log_printf(DEBUG, "Cleaned up bad device %s\n", port_name);
@@ -582,7 +582,7 @@ int receive_message(relay_t* relay, message_t* msg) {
     char port_name[MAX_PORT_NAME_SIZE];
     construct_port_name(port_name, relay->is_virtual, relay->is_usb, relay->port_num);
 
-    if (relay->dev_id.uid == -1) {
+    if (relay->dev_id.uid == (uint64_t) -1) {
         /* Haven't verified device is lowcar yet
          * read() is set to timeout while waiting for an ACK (see serialport_open())*/
         pthread_setcancelstate(PTHREAD_CANCEL_ENABLE, NULL);
@@ -872,7 +872,7 @@ int main(int argc, char* argv[]) {
     // If SIGINT (Ctrl+C) is received, call stop() to clean up
     signal(SIGINT, stop);
     init();
-	home_dir = getenv("HOME"); // set the home directory 
+    home_dir = getenv("HOME");  // set the home directory
     log_printf(INFO, "DEV_HANDLER initialized.");
     poll_connected_devices();
     return 0;
